% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GreedyStratifiedTeams.R
\name{GreedyStratifiedTeams}
\alias{GreedyStratifiedTeams}
\title{Greedy Algorithm Team Assignment - Accounting for Strata}
\usage{
GreedyStratifiedTeams(
  group_id,
  group_score,
  strata_score,
  num_players,
  num_strata,
  num_teams,
  num_groups,
  max_num_team
)
}
\arguments{
\item{group_id}{Identifier for group.}

\item{group_score}{Score associated with group members not in strata.}

\item{strata_score}{Score associated with strata.}

\item{num_players}{Number of players in each group.}

\item{num_strata}{Number of players in a specific strata in each group.}

\item{num_teams}{Number of teams to generate.}

\item{num_groups}{Number of groups included in group_id.}

\item{max_num_team}{Maximum number of groups / team.}
}
\description{
Greedy Algorithm Team Assignment - Accounting for Strata
}
