% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GreedyTeams.R
\name{GreedyTeams}
\alias{GreedyTeams}
\title{Greedy Algorithm Teams Assignment}
\usage{
GreedyTeams(
  group_id,
  group_score,
  num_players,
  num_teams,
  num_groups,
  max_num_team
)
}
\arguments{
\item{group_id}{Identifier for group.}

\item{group_score}{Score associated with group.}

\item{num_players}{Number of players in each group.}

\item{num_teams}{Number of teams to generate.}

\item{num_groups}{Number of groups included in group_id.}

\item{max_num_team}{Maximum number of groups / team.}
}
\description{
Greedy Algorithm Teams Assignment
}
