% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{mens_team}
\alias{mens_team}
\title{Small dataset for use in \pkg{rstapDP} examples and vignettes.}
\format{
A dataframe consisting of two columns: group_id and group_score.
}
\usage{
mens_team
}
\description{
Small dataset for use in \pkg{rstapDP} examples and vignettes.
}
\details{
A minimal dataset representing a typical mens league registration.
\describe{
\item{\code{group_id}}{The group (player or baggage group) unique identifier}
\item{\code{group_score}}{The score assigned to the group.}
}
}
\keyword{datasets}
