% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MILPTeams.R
\name{MILPTeams}
\alias{MILPTeams}
\title{Mixed Integer Linear Programming Teams Assignment}
\usage{
MILPTeams(group_id, group_score, num_teams, num_groups, max_num_team)
}
\arguments{
\item{group_id}{Identifier for group.}

\item{group_score}{Score associated with group.}

\item{num_teams}{Number of teams to generate.}

\item{num_groups}{Number of groups included in group_id.}

\item{max_num_team}{Maximum number of groups / team.}
}
\description{
Mixed Integer Linear Programming Teams Assignment
}
