% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateBalancedTeams.R
\name{GenerateBalancedTeams}
\alias{GenerateBalancedTeams}
\title{Generate Balanced Teams}
\usage{
GenerateBalancedTeams(
  group_id,
  group_score,
  num_teams,
  max_num_team = floor(length(group_id)/num_teams)
)
}
\arguments{
\item{group_id}{Identifier for group.}

\item{group_score}{Score associated with group.}

\item{num_teams}{Number of teams to generate.}

\item{max_num_team}{Maximum number of groups / team. Default value is set
to be the floor of the total number of groups divided by the number of
teams.}
}
\description{
Generate Balanced Teams
}
